requirements.txt
fastapi
uvicorn[standard]
pyvmomi
pyyaml
pydantic
cryptography
